RavenDB-24069 Consolidate BranchFallsBackToUnsharedAndRecoversCleanly into RavenDB_24069.cs
